scdiffusionX/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scdiffusionX"
version = "0.0.2"
authors = [
  { name="Erpai Luo", email="lep23@mails.tsinghua.edu.cn" },
]
description = "scDiffusion-X: Diffusion Model for Single-Cell Multiome Data Generation and Analysis"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  "blobfile>=2.0.0",
  "click>=8.1.7",
  "einops>=0.8.0",
  "hydra-core>=1.3.2",
  "hydra-submitit-launcher>=1.2.0",
  "hydra-optuna-sweeper>=1.2.0",
  "multiprocess>=0.70",
  "muon>=0.1.6",
  "numpy>=1.22.4",
  "omegaconf>=2.3.0",
  "pytorch_lightning<1.9.0",
  "PyYAML>=6.0.2",
  "rich>=13.9.4",
  "scanpy>=1.9.1",
  "scikit_learn>=1.2.2",
  "scipy>=1.10.1",
  "scvi-tools<=1.0.0",
  "tensorboard>=2.12.0",
  "torch>=2.1.2",
  "torchvision>=0.14.0",
  "tqdm>=4.64.1",
  "wandb>=0.16.1",
]

license = {file = "LICENSE"}
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/EperLuo/scDiffusion-X"
Issues = "https://github.com/EperLuo/scDiffusion-X/issues"